refactor(cli): break adapter import cycle with leaf contract module
The T3 s1/s2 consolidation put createExec/deriveLabelState and the
adapter types in deploy-status-adapter.ts, whose factory imports the
vendor adapters — which imported those helpers back, creating an
import/no-cycle violation (adapter.ts ↔ adapter-github/-linear/-jira)
caught only by the pre-push slow lint (the fast lint does not run
no-cycle, so the batch gates stayed green).

New leaf module deploy-status-adapter-shared.ts now holds the shared
surface — TrackerAdapter, CommentUpsertResult, ExecGh,
TrackerAdapterDeps, createExec, createExecGh, deriveLabelState — and
both the vendor adapters and the factory import FROM it one-way.
deploy-status-adapter.ts keeps the factory, TrackerAdapterOptions, and
withDryRun, and re-exports the moved surface verbatim so every
downstream importer (engine, cmd, refs, tests) is unchanged. JSDoc
moved intact; no behavior change (138 deploy-status tests green).


Work-Item: CodySwannGT/lisa#1968
